2-get-blog-posts
abstracted
